tout marche, mais affichage en doucle

+++ app/src/main/java/bts/sio/azurimmo/model/Appartement.kt
@@ -3,7 +3,7 @@ package bts.sio.azurimmo.model
 data class Appartement(
     val id: Int,
     val numero: String,
-    val surface: Double,
+    val surface: Float,
     val batiment: Batiment,
     val description: String
 )
